Add files via upload
This cybersecurity homelab project demonstrates a basic attack-defense scenario in a virtual environment. A Kali Linux machine is used to perform a TCP SYN scan on a Ubuntu machine using Nmap. Simultaneously, Wireshark is used on the Ubuntu machine to capture and analyze the network traffic in real-time. The project helps visualize how reconnaissance activities appear on a network and offers foundational hands-on experience in ethical hacking and network monitoring.